Add error checking for insertions
Apparently mini_s is some nullifying object (moreso than NULL or NA) that can't ever be used in if statements - not even length(mini_s) seems to work.

I cannot reproduce the issue on my machine, even with the same accession, so I'm skipping the bad insertions.
They're probably a result of NA insertions, which means don't insert, so skipping is the correct move.